Replaced select-none with no-select.
